chore: move configuration to separate class and application.yaml
