+++ src/main/java/com/hinderegger/steaminventorytracker/configuration/Constants.java
@@ -1,0 +1,9 @@
+package com.hinderegger.steaminventorytracker.configuration;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class Constants {
+    public static final String API_PATH = "/api/v1/items";
+}
