--- src/main/resources/application.properties
@@ -1,7 +1,0 @@
-spring.data.mongodb.authentication-database=admin
-spring.data.mongodb.username=rootuser
-spring.data.mongodb.password=rootpass
-spring.data.mongodb.database=steam-inventory-tracker
-spring.data.mongodb.port=27017
-spring.data.mongodb.host=localhost
-spring.data.mongodb.auto-index-creation=true
