+++ src/main/resources/application.yaml
@@ -1,0 +1,13 @@
+steam:
+  baseurl: "https://steamcommunity.com/market/priceoverview/?"
+  path: "appid=730&currency=3&market_hash_name="
+spring:
+  data:
+    mongodb:
+      authentication-database: admin
+      username: rootuser
+      password: rootpass
+      database: steam-inventory-tracker
+      port: 27017
+      host: localhost
+      auto-index-creation: true
